BuferGetFloat() requires a volatile pointer
For some reason, the returned data value is 0 unless *temp is marked
volatile. It seems like something is being optimized away.

Note that BuferGetFloat() seems to be called in only a single place
in StratoRatchuts and StratoRats, and that is in a log_nominal()
call in the MCBRouter. It's possible that this issue has never been
noticed before.